requirements.txt
Flask
Flask-SocketIO
Flask-SQLAlchemy
python-dotenv
google-generativeai
requests
eventlet
deepgram-sdk